+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 * add R7RS `guard` macro
 * add R7RS `parameterize` and `make-parameter`
+* add `shuffle` function
 ### Bugfix
 * remove evaluating of async list data as first argument
 * fix `number->string` for binary numbers